+++ util/Moire-lattice/CrI3/hdr/exchange.hpp
@@ -21,7 +21,7 @@
    extern double J_intra_3;
    extern double J_constant;
    extern double max_range;
-   extern int number_of_interactions;
+   extern uint64_t number_of_interactions;
    
    extern std::vector < std::vector < double > > Jint;
    extern std::vector < std::vector < double > > Jinter;
